+++ doc/Changelog.md
@@ -6,6 +6,7 @@
 
 * Fixed the location of temporary intermediate files for SoX-based audio effects
 * Enabled the "built-in" resampler for QAAC encoder
+* The "Algorithm Quality" slider now also effects the QAAC encoder
 * Updated MediaInfo to v0.7.82 (2016-01-27), compiled with ICL 15.0 and MSVC 12.0
 * Updated QAAC add-in to the to QAAC v2.58 (2016-01-05)
 * Updated ALAC decoder to refalac v1.58 (2016-01-05)